Introduce weapon attributes
Instead of fixed classes, like polearm, use a flexible attribute system.

All spears are polearms, but not all polearms are spears.  Some spears are two
handed, some are one handed, and some can be either.  Spears can have a blade
or be a point (think a needle).  Some spears can be thrown.  The butt of a spear
can be used for an attack.

The idea is to compose attributes to describe a weapon.  A javelin might be a
blade and range weapon.  A lance might be a blade and two handed weapon.  Both
might be considered a spear.

Partially addresses, partially deviates from #15.